doc: place generated documents in the doc_gen folder
added a new facet that will place documents generated into the "doc_gen" folder instead of "source_gen".
Also refactored the HTML generator to generate a single file per document. This doesn't work wit the latex
generator yet. I was required to move the HTMLRenderer into the document configuration instead of the
ExportConfiguration. The overall goal is to make the documentation language more flexable and allow to
generate multiple or a single file from a set of documents.